+++ Options/Defaults.lua
@@ -91,6 +91,7 @@ addon.Options.Defaults = {
             TabPriority = 5,
             ShowWarbandIcon = true,
             WarbandHeaderColor = true,
+            ShowOtherFactionWarbandAsCompleted = true
         },
         RightClickMenu = {
             ShowButtonOnAchievement = false,