+++ Options/Layout.lua
@@ -1058,7 +1058,18 @@ local achievementsOptions = {
                             get = function() return addon.Options.db.profile.Achievements.HideDateCompleted; end,
                             set = function(_, value) addon.Options.db.profile.Achievements.HideDateCompleted = value; end,
                         },
-                        Blank1 = {order = OrderPP(), type = "description", width = AdjustedWidth(1.35), name = ""},
+                        ShowOtherFactionWarbandAsCompleted = {
+                            order = OrderPP(), type = "toggle", width = AdjustedWidth(1.35),
+                            name = addon.L["Show Other Faction Warband as Completed"],
+                            desc = function() return addon.L["Show Other Faction Warband as Completed Desc"]:K_ReplaceVars{
+                                    hordeIntro = addon.GetAchievementInfoTable(12555).Name,
+                                    allianceIntro = addon.GetAchievementInfoTable(12582).Name,
+                                    warStories = addon.GetAchievementInfoTable(40955).Name,
+                                }:KAF_AddDefaultValueText("Achievements.ShowOtherFactionWarbandAsCompleted")
+                            end,
+                            get = function() return addon.Options.db.profile.Achievements.ShowOtherFactionWarbandAsCompleted; end,
+                            set = function(_, value) addon.Options.db.profile.Achievements.ShowOtherFactionWarbandAsCompleted = value; end,
+                        },
                         ShowAllianceFactionIcon = {
                             order = OrderPP(), type = "toggle", width = AdjustedWidth(1.35),
                             name = addon.L["Show Faction Faction Icon"]:K_ReplaceVars(addon.L["Alliance"]),